XML configuration line marker improvements
- Each type of XmlNodes now has custom user-friendly representation, when rendered in line marker
- Corresponding Magento area is specified for each item in line marker
